change uint to uint256

+++ token/ERC20Basic.sol
@@ -7,8 +7,8 @@ pragma solidity ^0.4.8;
  * @dev see https://github.com/ethereum/EIPs/issues/20
  */
 contract ERC20Basic {
-  uint public totalSupply;
-  function balanceOf(address who) constant returns (uint);
-  function transfer(address to, uint value) returns(bool success);
-  event Transfer(address indexed from, address indexed to, uint value);
+  uint256 public totalSupply;
+  function balanceOf(address who) constant returns (uint256);
+  function transfer(address to, uint256 value) returns(bool success);
+  event Transfer(address indexed from, address indexed to, uint256 value);
 }
